Fix issue with CosmosAsync and Spotify Web API
