+++ libs/shared/src/platform/clipboard.ts
@@ -1,0 +1,4 @@
+export type ClipboardAPI = {
+    copy(value: any): Promise<void>;
+    paste(): Promise<string>;
+};
