+++ libs/shared/src/platform/session.ts
@@ -1,0 +1,8 @@
+export type Session = {
+    accessToken: string;
+    accessTokenExpirationTimestampMs: number;
+    isAnonymous: boolean;
+    locale: string;
+    market: string;
+    valid: boolean;
+};
