+++ libs/spotify-web-api/src/index.ts
@@ -1,2 +1,3 @@
 export * from './api';
 export * from './models';
+export * from './utils';
