+++ libs/spotify-web-api/tsconfig.json
@@ -1,4 +1,5 @@
 {
+    "extends": "../../tsconfig.base.json",
     "compilerOptions": {
         "target": "ES2017",
         "module": "commonjs",
@@ -7,5 +8,6 @@
         "forceConsistentCasingInFileNames": true,
         "strict": true,
         "skipLibCheck": true
-    }
+    },
+    "include": ["./src/**/*.*"]
 }
